+++ frontend/src/App.vue
@@ -1,16 +1,14 @@
 <script setup>
 import Wizzard from "./components/Wizzard.vue";
-import Dashboard from "./components/Dashboard.vue";
 import RebuildFrontend from "./components/RebuildFrontend.vue";
 
 import {ref} from "vue";
-import imgUrl from './assets/banner.jpg'
+import imgUrl from './assets/banner.png'
 
 const develop = ref(false);
 const choosing = ref(true);
 
 const usingWizzard = ref(false);
-const usingDashboard = ref(false);
 const usingRebuildFrontend = ref(false);
 
 const chooseWizzard = () => {
@@ -18,10 +16,6 @@ const chooseWizzard = () => {
     choosing.value = false;
 }
 
-const chooseDashboard = () => {
-    usingDashboard.value = true;
-    choosing.value = false;
-}
 
 const chooseRebuildFrontend = () => {
     usingRebuildFrontend.value = true;
@@ -31,28 +25,49 @@ const chooseRebuildFrontend = () => {
 <template>
     <div>
         <div v-if="choosing">
-            <header class="p-5 mb-5 bg-gray-700 text-white text-left">
-                <h1 class="navbar-brand">Repository Generator</h1>
+            <header class="p-5 mb-5 bg-[#272822] text-white text-left">
+                <h1 class="navbar-brand">
+                    <span class="font-bold text-[#F92672]">class</span>
+                    Repository
+                    <span class="font-bold text-[#F92672]">extends</span>
+                    Generator
+                </h1>
             </header>
             <div class="grid grid-cols-12 gap-4 px-5">
                 <div class="col-span-6">
                     <img :src="imgUrl" alt="Repository Generator Factory" class="banner">
                 </div>
                 <div class="col-span-6">
-                    <button v-if="develop" class="disabled:bg-slate-50 disabled:text-slate-500 disabled:border-slate-200 disabled:shadow-none block w-full py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-gray-400 hover:bg-gray-500"
-                            @click="chooseDashboard">
-                        Dashboard
-                    </button>
+                    <div class="code-menu hide-scrollbar  mt-[50vw] h-auto w-full overflow-auto rounded-xl bg-[#272822] px-3 py-6 font-bold text-white lg:relative lg:right-0 lg:ml-auto lg:mt-[80px] lg:w-[385px] xl:right-[5%] 2xl:right-[10%] widescreen:right-[18%]" >
+                        <div class="font-bold">
+                            <span class="font-bold text-[#F92672]">protected </span><span class="font-bold text-white">$menu </span>
+                            <span class="font-bold text-[#F92672]">=</span> [
+                            <a class="code-link relative my-2 block rounded border-none pl-4 text-left text-blue hover:bg-blue hover:text-white" style="margin-left: 10px; width: calc(100% - 10px);" href="#" @click="chooseWizzard">
+                                <div class="text-[.7rem] font-normal text-white/50">// lets start</div>
+                                <span class="font-bold text-[#E6DB74]">"generate" </span>
+                                <span class="font-bold text-[#F92672]">=&gt;</span>
+                                <span class="font-bold text-[#E6DB74]">"New Domain"</span>
+                                <span class="text-white">,</span>
+                            </a>
 
-                    <button class="mt-3 block w-full py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700"
-                            @click="chooseWizzard">
-                        Wizzard
-                    </button>
+                            <a class="code-link relative my-2 block rounded border-none pl-4 text-left text-blue hover:bg-blue hover:text-white" style="margin-left: 10px; width: calc(100% - 10px);" href="#" @click="chooseRebuildFrontend">
+                                <div class="text-[.7rem] font-normal text-white/50">// if need to fix</div>
+                                <span class="font-bold text-[#E6DB74]">"re_generate" </span>
+                                <span class="font-bold text-[#F92672]">=&gt;</span>
+                                <span class="font-bold text-[#E6DB74]">"Old Domain"</span>
+                                <span class="text-white">,</span>
+                            </a>
 
-                    <button class="mt-3 block w-full py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700"
-                            @click="chooseRebuildFrontend">
-                        Rebuild frontend
-                    </button>
+                            <a class="code-link relative my-2 block rounded border-none pl-4 text-left text-blue hover:bg-blue hover:text-white" style="margin-left: 10px; width: calc(100% - 10px);" href="#">
+                                <div class="text-[.7rem] font-normal text-white/50">// read and learn</div>
+                                <span class="font-bold text-[#E6DB74]">"help" </span>
+                                <span class="font-bold text-[#F92672]">=&gt;</span>
+                                <span class="font-bold text-[#E6DB74]">"How it works"</span>
+                                <span class="text-white">,</span>
+                            </a>
+                            <span class="font-bold">],</span>
+                        </div>
+                    </div>
                 </div>
             </div>
         </div>
@@ -60,7 +75,6 @@ const chooseRebuildFrontend = () => {
             <Wizzard v-if="usingWizzard"
                      class="bg-white p-10 shadow max-w-screen-2xl mx-auto py-12 px-4 sm:px-6 lg:px-8"/>
             <RebuildFrontend v-else-if="usingRebuildFrontend" class="bg-white p-10 shadow max-w-screen-2xl mx-auto py-12 px-4 sm:px-6 lg:px-8" />
-            <Dashboard v-else-if="usingDashboard"/>
         </div>
     </div>
 </template>