+++ frontend/src/components/Step1.vue
@@ -1,7 +1,8 @@
 <script setup>
-import {defineEmits, ref} from 'vue'
+import {defineEmits, ref, computed} from 'vue'
+import axios from "axios";
 
-const emit = defineEmits(['nameChanged', 'iconChanged'])
+const emit = defineEmits(['nameChanged', 'iconChanged', 'tableExists'])
 const props = defineProps({
     modelName: {
         type: String,
@@ -16,6 +17,28 @@ const props = defineProps({
 })
 let name = ref(props.modelName);
 let icon = ref(props.icon);
+
+const tableExists = ref(false);
+
+const checkTable = () => {
+    emit('nameChanged',name.value)
+
+    axios.get(import.meta.env.VITE_API_URL + '/repgenerator/isTableExist/'+name.value).then((response) => {
+        tableExists.value = response.data;
+
+        emit('tableExists', tableExists)
+    })
+}
+
+const ucfirstName = computed({
+    get() {
+        return name.value;
+    },
+    set(ucName) {
+        if(ucName.length < 1) {name.value = ''; return}
+        name.value = ucName.replace(/^./, ucName[0].toUpperCase());
+    }
+})
 </script>
 
 <template>
@@ -25,8 +48,9 @@ let icon = ref(props.icon);
                 <label class="block text-sm font-medium text-gray-700" for="model-name">
                     Model Name (Singular - Ex. Dog )
                 </label>
+                <p v-if="tableExists" class="mt-2 text-sm text-red-600" id="table-error">This table name is already exists</p>
                 <div class="mt-1">
-                    <input id="model-name" v-model="name" class="shadow-sm block w-full sm:text-sm border-gray-300 rounded-md" required type="text" @change="emit('nameChanged',name)">
+                    <input id="model-name" v-focus v-model="ucfirstName" class="shadow-sm block w-full sm:text-sm  rounded-md" :class="tableExists ? 'border-red-300 text-red-900 placeholder-red-300 focus:outline-none focus:ring-red-500 focus:border-red-500' : 'border-gray-300' " required type="text" @change="checkTable">
                 </div>
                 <label class="block text-sm font-medium text-gray-700 mt-1" for="model-name">
                     Hero icon <a target="_blank" href="https://heroicons.com/">https://heroicons.com/</a>