+++ frontend/src/components/Wizzard.vue
@@ -8,9 +8,9 @@ import axios from 'axios'
 import Options from "./Options.vue";
 import factoryImgUrl from '../assets/factory.gif'
 
-
 // Wizzard
 const stepNumber = ref(1);
+
 const steps = [
     {
         title: 'Name'
@@ -25,6 +25,7 @@ const steps = [
         title: 'Result'
     }
 ];
+
 const step1Options = ref({
     'pivot': {
         label: 'Pivot',
@@ -38,13 +39,20 @@ const step1Options = ref({
     },
     'softDelete': {
         label: 'Soft Delete',
-        enabled: false,
+        enabled: true,
         text: 'Is the model use soft delete?'
+    },
+    'timestamps': {
+        label: 'Timestamps',
+        enabled: true,
+        text: 'Is the model use timestamps?'
     }
 });
+
 const scrollToTop = () => {
     document.getElementById('scroll-anchor').scrollIntoView({behavior: "smooth", block: "end", inline: "nearest"});
 }
+
 const resetWizzard = () => {
     stepNumber.value = 1;
     modelName.value = '';
@@ -54,8 +62,11 @@ const resetWizzard = () => {
         step1Options[index].enabled = false;
     }
 }
+
 const messages = ref([]);
+
 const generating = ref(false);
+
 const generate = () => {
     generating.value = true;
     let payload = {
@@ -75,6 +86,7 @@ const generate = () => {
         generating.value = false;
     })
 }
+
 const onNextStep = (e) => {
     e.preventDefault();
     if (stepNumber.value === steps.length) {
@@ -86,26 +98,33 @@ const onNextStep = (e) => {
         ++stepNumber.value;
     }
 }
+
 const onPreviousStep = (e) => {
     e.preventDefault();
     stepNumber.value = stepNumber.value - 1;
 }
+
 const isPreviousDisabled = () => {
     return stepNumber.value <= 1;
 }
+
 const isOverview = () => {
     return stepNumber.value === steps.length - 1;
 }
+
 const isLastStep = () => {
     return stepNumber.value === steps.length;
 }
 
 // 1. Name
 const modelName = ref('');
+
 const icon = ref('');
+
 const onNameChanged = (name) => {
     modelName.value = name;
 }
+
 const onIconChanged = (setIcon) => {
     icon.value = setIcon;
 }
@@ -136,58 +155,11 @@ const getDefaultColumns = () => {
             'is_picture': false,
             'is_hashed': false,
             'is_crypted': false,
-        },
-        {
-            'name': 'created_at',
-            'type': 'timestamp',
-            'length': '',
-            'precision': '',
-            'scale': '',
-            'default': '',
-            'auto_increment': false,
-            'nullable': false,
-            'reference': '',
-            'foreign': '',
-            'cascade': '',
-            'searchable': '',
-            'values': '',
-            'comment': '',
-            'unsigned': false,
-            'index': [],
-            'show_on_table': false,
-            'uploads_files_path': '',
-            'is_file': false,
-            'is_picture': false,
-            'is_hashed': false,
-            'is_crypted': false,
-        },
-        {
-            'name': 'updated_at',
-            'type': 'timestamp',
-            'length': '',
-            'precision': '',
-            'scale': '',
-            'default': '',
-            'auto_increment': false,
-            'nullable': false,
-            'reference': '',
-            'foreign': '',
-            'cascade': '',
-            'searchable': '',
-            'values': '',
-            'comment': '',
-            'unsigned': false,
-            'index': [],
-            'show_on_table': false,
-            'uploads_files_path': '',
-            'is_file': false,
-            'is_picture': false,
-            'is_hashed': false,
-            'is_crypted': false,
         }
     ]
 }
 const columns = ref(getDefaultColumns());
+
 const onAddColumn = () => {
     columns.value.push({
         'name': '',
@@ -222,18 +194,40 @@ const onRemoveColumn = (data) => {
         }
     }
 }
+
+const onRefreshTables = () => {
+    getTables()
+}
+
+const isTableExists = ref(false);
+
+const onTableExists = (data) => {
+    isTableExists.value = data.value;
+}
+
+const onSelectOption = (data) => {
+    if(data.label === "Pivot") {
+
+    }
+}
+
 // 2. Models
 const models = ref([]);
-axios.get(import.meta.env.VITE_API_URL + '/repgenerator/tables').then((response) => {
-    models.value = response.data;
-})
+
+const getTables = () => {
+    axios.get(import.meta.env.VITE_API_URL + '/repgenerator/tables').then((response) => {
+        models.value = response.data;
+    })
+}
+
+getTables()
 </script>
 
 <template>
     <div class="grid place-items-center h-screen" v-if="generating">
         <img :src="factoryImgUrl" alt="Repository Generator Factory" class="factory">
     </div>
-    <form @submit="onNextStep" v-else>
+    <form class="mt-5" @submit="onNextStep" v-else>
         <nav aria-label="Progress">
             <ol class="border border-gray-300 rounded-md divide-y divide-gray-300 md:flex md:divide-y-0" role="list">
                 <Step v-for="(step,index) in steps" :complete="index+1 < stepNumber" :current="stepNumber === index+1"
@@ -241,11 +235,11 @@ axios.get(import.meta.env.VITE_API_URL + '/repgenerator/tables').then((response)
             </ol>
         </nav>
         <div v-if="stepNumber === 1 || isOverview()">
-            <Step1 :icon="icon" :modelName="modelName" @iconChanged="onIconChanged" @nameChanged="onNameChanged"/>
-            <Options :options="step1Options"/>
+            <Step1 :icon="icon" :modelName="modelName"  @iconChanged="onIconChanged" @nameChanged="onNameChanged" @tableExists="onTableExists"/>
+            <Options :options="step1Options" @selectOption="onSelectOption"/>
         </div>
         <Step2 v-if="stepNumber === 2 || isOverview()" :columns="columns" :disableAdd="isOverview()" :models="models" :modelName="modelName"
-               @addColumn="onAddColumn" @removeColumn="onRemoveColumn"/>
+               @addColumn="onAddColumn" @removeColumn="onRemoveColumn" @refreshTables="onRefreshTables"/>
         <Result v-if="isLastStep()" :messages="messages"/>
 
         <div v-if="!isLastStep()" class="pt-5 grid grid-cols-12 gap-4">
@@ -259,7 +253,8 @@ axios.get(import.meta.env.VITE_API_URL + '/repgenerator/tables').then((response)
             </div>
             <div class="col-span-6">
                 <button
-                    class="block w-full py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700"
+                    :disabled="isTableExists"
+                    class="disabled:bg-slate-50 disabled:text-slate-500 disabled:border-slate-200 disabled:shadow-none block w-full py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700"
                     type="submit">
                     {{ isOverview() ? 'Finish' : 'Next' }}
                 </button>